+++ Source/lib/embeddedLapack/src/Lapack/F2c/r_sign.c
@@ -4,12 +4,12 @@ extern "C" {
 #endif
 
 #ifdef KR_headers
-double r_sign(a,b) real *a, *b;
+float r_sign(a,b) real *a, *b;
 #else
-double r_sign(real *a, real *b)
+float r_sign(real *a, real *b)
 #endif
 {
-double x;
+float x;
 x = (*a >= 0 ? *a : - *a);
 return( *b >= 0 ? x : -x);
 }
