+++ Source/lib/embeddedLapack/src/LinearAlgebra/diagpower.c
@@ -11,7 +11,7 @@
  * Take power p, on all element on the diagonal of matrix A, who has the size row x column
  */
 
-void diagpower(double* A, double p, int row, int column){
+void diagpower(float* A, float p, int row, int column){
 
 	for(int i = 0; i < row; i++){
 		for(int j = 0; j < column; j++){
