+++ Source/lib/embeddedLapack/src/LinearAlgebra/print.c
@@ -9,7 +9,7 @@
 /*
  * Print a matrix A, with the dimension row x column
  */
-void print(double* A, int row, int column) {
+void print(float* A, int row, int column) {
 	for(int i = 0; i < row; i++){
 		for(int j = 0; j < column; j++){
 			printf("%0.18f ", *(A++));
