+++ Source/lib/embeddedLapack/src/LinearAlgebra/quadprog.c
@@ -25,7 +25,7 @@
  *  Set SHOW_QP_OUTPUT to TRUE to begin with in declareFunctions.h
  *  I recommend to read the qpOASES_e manual
  */
-void quadprog(double* H, double* g, double* A, double* ulb_vec, double* uub_vec, double* ylb_vec, double* yub_vec, int* nWSR, double* u, int columnH, int rowA) {
+void quadprog(float* H, float* g, float* A, float* ulb_vec, float* uub_vec, float* ylb_vec, float* yub_vec, int* nWSR, float* u, int columnH, int rowA) {
 
 	// Setting up QProblem object J = U^T*H*U + g^T*U
 	QProblem objective;
