+++ Source/lib/embeddedLapack/src/LinearAlgebra/sqrte.c
@@ -11,7 +11,7 @@
 /*
  * Take the square root of every element of matrix A, size row x column
  */
-void sqrte(double* A, int row, int column){
+void sqrte(float* A, int row, int column){
 	for(int i = 0; i < row; i++)
 		for(int j = 0; j < column; j++){
 			*A = sqrt(*(A));
