+++ Source/lib/embeddedLapack/src/LinearAlgebra/sub.c
@@ -11,7 +11,7 @@
  * Sub C = A-B with A size row_a x column_a and B have the size row_a x column_b and C have the size row_a x column_a
  */
 
-void sub(double* A, double* B, double* C, int row_a, int column_a, int column_b) {
+void sub(float* A, float* B, float* C, int row_a, int column_a, int column_b) {
 
 	/*
 	 * This uses the formula C=A-B
